+++ engine/include/analyzer/modules/string/string_analyzer.h
@@ -1,0 +1,10 @@
+/* String Analyzer Submodule Header */
+
+#ifndef VMCU_STRING_ANALYZER_H
+#define VMCU_STRING_ANALYZER_H
+
+typedef struct vmcu_report vmcu_report_t;
+
+extern int vmcu_analyze_strings(vmcu_report_t *report);
+
+#endif